+++ tomcat/src/main/java/org/apache/catalina/servlets/Container.java
@@ -1,9 +1,8 @@
-package org.apache.catalina.container;
+package org.apache.catalina.servlets;
 
 import common.FileReader;
 import nextstep.mvc.ResponseWriter;
-import org.apache.catalina.controller.Controller;
-import org.apache.catalina.controller.config.ControllerMapping;
+import org.apache.catalina.servlets.config.ServletMapping;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.apache.coyote.http11.HttpStatus;
@@ -11,11 +10,12 @@ import org.apache.coyote.http11.HttpStatus;
 public class Container {
 
     private Container() {
+
     }
 
     public static void process(HttpRequest request, HttpResponse response) {
-        Controller controller = ControllerMapping.getController(request);
-        if (controller == null) {
+        Servlet servlet = ServletMapping.getServlet(request);
+        if (servlet == null) {
             if (FileReader.findResource(request.getPath()) == null) {
                 ResponseWriter.view(response, HttpStatus.NOT_FOUND, "/404.html");
                 return;
@@ -23,6 +23,6 @@ public class Container {
             ResponseWriter.view(response, HttpStatus.OK, request.getPath());
             return;
         }
-        controller.service(request, response);
+        servlet.service(request, response);
     }
 }
