+++ tomcat/src/main/java/org/apache/catalina/servlets/HelloServlet.java
@@ -1,18 +1,18 @@
-package org.apache.catalina.controller;
+package org.apache.catalina.servlets;
 
 import nextstep.mvc.ResponseWriter;
-import org.apache.catalina.controller.config.RequestMapping;
-import org.apache.catalina.controller.exception.MethodNotAllowedException;
+import org.apache.catalina.servlets.config.RequestMapping;
+import org.apache.catalina.servlets.exception.MethodNotAllowedException;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.apache.coyote.http11.HttpStatus;
 
-@RequestMapping("/index")
-public class IndexController extends AbstractController {
+@RequestMapping({"/", ""})
+public class HelloServlet extends HttpServlet {
 
     @Override
     protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
-        ResponseWriter.view(httpResponse, HttpStatus.OK, httpRequest.getUri());
+        ResponseWriter.writeWithData(httpResponse, HttpStatus.OK, "Hello world!");
     }
 
     @Override
