+++ tomcat/src/main/java/org/apache/catalina/servlets/Servlet.java
@@ -1,9 +1,9 @@
-package org.apache.catalina.controller;
+package org.apache.catalina.servlets;
 
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
-public interface Controller {
+public interface Servlet {
 
     void service(HttpRequest request, HttpResponse response);
 }
