+++ tomcat/src/main/java/org/apache/catalina/servlets/config/ServletMapping.java
@@ -1,30 +1,30 @@
-package org.apache.catalina.controller.config;
+package org.apache.catalina.servlets.config;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import org.apache.catalina.controller.Controller;
+import org.apache.catalina.servlets.Servlet;
 import org.apache.coyote.http11.HttpRequest;
 import org.reflections.Reflections;
 
-public class ControllerMapping {
+public class ServletMapping {
 
-    private static final Map<String, Controller> controllers = new ConcurrentHashMap<>();
+    private static final Map<String, Servlet> servlets = new ConcurrentHashMap<>();
 
-    private ControllerMapping() {
+    private ServletMapping() {
     }
 
     public static void initialize() {
-        Reflections reflections = new Reflections("org.apache.catalina.controller");
+        Reflections reflections = new Reflections("org.apache.catalina.servlets");
         Set<Class<?>> declaredClasses = reflections.getTypesAnnotatedWith(RequestMapping.class);
         for (final Class<?> clazz : declaredClasses) {
             RequestMapping annotation = clazz.getAnnotation(RequestMapping.class);
             String[] uris = annotation.value();
             try {
-                Controller controller = (Controller) clazz.getConstructor().newInstance();
+                Servlet servlet = (Servlet) clazz.getConstructor().newInstance();
                 for (final String uri : uris) {
-                    controllers.put(uri, controller);
+                    servlets.put(uri, servlet);
                 }
             } catch (NoSuchMethodException | SecurityException | IllegalAccessException |
                      InstantiationException |
@@ -34,11 +34,11 @@ public class ControllerMapping {
         }
     }
 
-    public static Controller getController(HttpRequest request) {
-        Controller controller = controllers.get(request.getPath());
-        if (controller == null) {
-            return controllers.get(request.getPath() + ".html");
+    public static Servlet getServlet(HttpRequest request) {
+        Servlet servlet = servlets.get(request.getPath());
+        if (servlet == null) {
+            return servlets.get(request.getPath() + ".html");
         }
-        return controller;
+        return servlet;
     }
 }
