+++ tomcat/src/main/java/org/apache/catalina/servlets/exception/MethodNotAllowedException.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.controller.exception;
+package org.apache.catalina.servlets.exception;
 
 public class MethodNotAllowedException extends RuntimeException {
